makefile upd, google formatted code style, qt directory rename
