+++ Makefile
@@ -1,9 +1,9 @@
 UNAME = $(shell uname -s)
 FLAGSS = "noflags"
 CC = gcc -Wall -Werror -Wextra -std=c11 -g
-SOURCES_C = src/engine/s21_camera.c src/engine/s21_camera.h src/engine/s21_engine.h src/engine/s21_mesh.c src/engine/s21_mesh.h src/engine/math/s21_engine_math.c src/engine/math/s21_engine_math.h src/engine/math/s21_matrix4x4.c src/engine/math/s21_matrix4x4.h
-SOURCES_CPP = "none"
-SOURCES_UI = "none"
+SOURCES_C = src/engine/s21_engine.h src/engine/s21_mesh.c src/engine/s21_mesh.h src/engine/math/s21_engine_math.c src/engine/math/s21_engine_math.h src/engine/math/s21_matrix4x4.c src/engine/math/s21_matrix4x4.h
+SOURCES_CPP = src/main.cpp src/gui/mainwindow.cpp src/gui/mainwindow.h src/gui/openglwidget.cpp src/gui/openglwidget.h
+SOURCES_UI = src/gui/mainwindow.ui
 SOURCES_COMPILED = "none"
 SOURCES_TESTS = "none"
 TMPDIR = tmp
@@ -22,12 +22,8 @@ ifeq ($(UNAME),Linux)
 	FLAGSS = -lcheck -lsubunit -lrt -lm -lpthread -fprofile-arcs
 endif
 
-all: base
+all: cleanall install dist dvi launch
 .PHONY : all
-
-base: $(SOURCES_C)
-	gcc -Wall -Werror -Wextra src/engine/s21_camera.c src/engine/s21_mesh.c src/engine/math/s21_matrix4x4.c src/engine/math/s21_engine_math.c -lm -lglut -lGL -lGLU -o app -g
-.PHONY : base
 
 build: $(SOURCES_C) $(SOURCES_CPP) $(SOURCES_UI)
 	if [ ! -d "$(BUILDDIR_RELEASE)" ] && [ ! -f "build/$(OUTNAME)" ]; then cd build/ && make; fi
